renamed task_actions to task_http_actions because I am going to implement real time socket update, also added a lot of documentation comments to my shits
