+++ frontend/components/projects/project_editor.jsx
@@ -1,6 +1,8 @@
+// React
 import React from 'react';
 import PropTypes from 'prop-types';
 import ReactQuill from 'react-quill';
+// Material UI
 import TextField from 'material-ui/TextField';
 import FlatButton from 'material-ui/FlatButton';
 import DatePicker from 'material-ui/DatePicker';
