+++ frontend/components/projects/project_task_item.jsx
@@ -1,3 +1,4 @@
+// React
 import React from 'react';
 import PropTypes from 'prop-types';
 import Paper from 'material-ui/Paper';
