+++ frontend/containers/projects/project_detail.jsx
@@ -1,19 +1,21 @@
+// React
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-
-
-import LinearProgress from 'material-ui/LinearProgress';
-import Chip from 'material-ui/Chip';
+// Material UI
 import { PieChart, Pie, Tooltip, Cell } from 'recharts';
+import Chip from 'material-ui/Chip';
+import LinearProgress from 'material-ui/LinearProgress';
 import CircularProgress from 'material-ui/CircularProgress';
-
 // Actions
 import { fetchProjectDetail } from '../../actions/project_actions';
 import ProjectTaskItem from '../../components/projects/project_task_item';
-
+// Enums
 const COLORS = ['#00C49F', '#0088FE', '#ff2828'];
-
+const circuleProgressStyle = {
+  top: '50%',
+  left: '50%'
+};
 
 class ProjectDetail extends React.Component {
   state = {
@@ -31,8 +33,17 @@ class ProjectDetail extends React.Component {
     this.props.dispatchFetchProjectDetail(this.props.match.params.id);
   }
 
+  /*
+   * Whenever component is going to unmount, always check if there is a callback waiting to be called. If so, cancel it.
+   */
+  componentWillUnmount() {
+    if (this.setTimeoutId) {
+      clearTimeout(this.setTimeoutId);
+    }
+  }
+
   /**
-   * Checks whether the project is fetched from the backend
+   * Checks whether the project is fetched from the backend.
    * @returns {boolean}
    */
   get isProjectLoading() {
@@ -40,6 +51,10 @@ class ProjectDetail extends React.Component {
     return this.props.project.id !== routerProjectId;
   }
 
+  /**
+   * Returns a loading bar that indicates to users whether the data are fetched from backend.
+   * @returns {React.Element}
+   */
   get progressIndicator() {
     if (this.isProjectLoading) {
       return <LinearProgress mode={'indeterminate'} />;
@@ -48,6 +63,11 @@ class ProjectDetail extends React.Component {
     return  <LinearProgress mode={'determinate'} value={100} />;
   }
 
+  /**
+   * Returns a project summary component which is really a rectangle that has two inner rectangles in it. The left rectangle
+   * is the project description along with people that are assigned to the project. The right rectangle is a pie chart.
+   * @returns {React.Element}
+   */
   get projectSummary() {
     if (this.isProjectLoading) {
       return;
@@ -68,8 +88,11 @@ class ProjectDetail extends React.Component {
     );
   }
 
+  /**
+   * Returns a pie chart which indicates to user the state of the project.
+   * @returns {React.Element}
+   */
   get pieChart() {
-    // NOTE: When page is refreshed, the pie chart fails to render because the height and width is null
     const containerHeight = $('#chart-container').height();
     const containerWidth = $('#chart-container').width();
 
@@ -87,7 +110,6 @@ class ProjectDetail extends React.Component {
         { name: 'Overdue', value: tasks.filter((task) =>  isOverDue(task)).length }
       ];
 
-      console.log(this.props.project.tasks, 'rendering pie chart');
       return (
         <PieChart width={$('#chart-container').width()} height={$('#chart-container').height()}>
           <Pie
@@ -104,11 +126,22 @@ class ProjectDetail extends React.Component {
         </PieChart>
       );
     }
-    setTimeout(() => {
-      this.setState({ height: containerHeight, width: containerWidth });
-    }, 500);
+    // To keep shit visually pleasing, it is actually BETTER to make user wait longer for the chart to show up because
+    // the circular progress bar should at least complete one cycle of spinning.
+    this.setTimeoutId = setTimeout(() => {
+      this.setState({
+        height: $('#chart-container').height(),
+        width: $('#chart-container').width()
+      });
+    }, 1000);
+
+    return <CircularProgress style={circuleProgressStyle} />;
   }
 
+  /**
+   * Returns a simple rectangular text container which has name of the project and description of the project in it.
+   * @returns {React.Element}
+   */
   get projectDescription() {
     return (
       <div className="project-description">
@@ -118,6 +151,10 @@ class ProjectDetail extends React.Component {
     );
   }
 
+  /**
+   * Returns a list of people who are assigned to the project.
+   * @returns {React.Element}
+   */
   get assigneeSection() {
     if (this.isProjectLoading) {
       return;
@@ -148,6 +185,10 @@ class ProjectDetail extends React.Component {
     );
   }
 
+  /**
+   * Returns a list of project task items contained within a task summary rectangle.
+   * @returns {React.Element}
+   */
   get taskSummary() {
     if (this.isProjectLoading) {
       return;
